[i915] Fix UB related to 0-length VLA
In C and C++, a VLA must have a positive, non-zero length. When
`CallOnDisplaysChanged` is invoked with no displays added and only
displays removed, the two VLAs in this function are declared to have
zero length, which invokes undefined behavior. Importantly, this
is a valid invocation for this function.

Bug: 96550
Change-Id: Ic5386fe5a2834a0b6f50141baf119f5d191c0656
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663782
Reviewed-by: Randall Bosetti <rlb@google.com>
Commit-Queue: Devon H. O'Dell <dhobsd@google.com>