[ffx] Stop all instances of emu on error
When using the '--all' flag to stop multiple emulators, if `ffx emu`
hits an error, it terminates the process early. This can leave some
emulators running, when all of the running emulators should be shut
down.

This CL adjusts the error handling code so that the errors it encounters
are printed to the screen, but it doesn't stop processing. Now, even if
multiple emulators are "broken", `ffx emu` will attempt to stop them
all.

Bug: fxbug.dev/95935
Change-Id: I748adecbe280579e36f8f6fe6486e3b39204bf43
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663482
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>