[test_manager] Propagate channel closure in mock
I noticed that the tests that are hanging in ffx test also access their
own diagnostics. This is supported in test framework via mocks that
forward connections to archivist and make minor modifications to the
output.
It's possible for a client to make a request, and then terminate before
reading the result. At the same time, if archivist doesn't have any
data to send back, the mock might wait forever and keep archivist alive
during shutdown. This eliminates this case by simultaneously polling
for client's channel to close and terminating the mock, even if it's
waiting for a response from archivist.

In theory, this could help eliminate some hangs in ffx test.

Testing: fx test //src/sys/test_manager
Bug: 94333

Change-Id: Ib53e863d49101ffda1ce0caa7172cb4f913b3064
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663113
Reviewed-by: Ankur Mittal <anmittal@google.com>
Commit-Queue: Satsuki Ueno <satsukiu@google.com>