[wlan] Create network device library components
Create NetworkDevice and NetworkPort classes to make it easier to
implement support for the fuchsia.hardware.network.device protocol for
an improved data plane in wlan drivers.

This consists of two classes that a driver writer can use to represent a
device and one or more ports associated with that device. These classes
intend to help the driver writer working with network devices and
streamline the work specifically for wlan drivers.

Test: Included unit tests, separate experimental tests
Bug: 79622
Change-Id: I888ea18a0af924bfd2ef87b03ae80ce2a6a34b6a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/567710
Reviewed-by: Sakthi Vignesh Radhakrishnan <rsakthi@google.com>
Commit-Queue: Bjoern Johansson <bjoernj@google.com>