[kernel][vm] Lookup gives offset on original object
Changes Lookup, and related methods that callback to a
VmObject::LookupFunction, to normalize the offset to be in terms of the
original object.

This has no impact to any existing callers, as all existing callers are
agnostic of whether the offsets were shifted or not due to slices, but
this provides a more consistent interface.

Change-Id: I4a58057076b881ecf28e133df30cac3aee4cf7b2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/662640
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>