[starnix] Run all tests with #[fuchsia::test]
This macro sets up logging, so failing tests are easier to debug.
Note that this macro also works for non-async tests.

Due to our use of glob imports, #[::fuchsia::test] needs to be used
(note the leading `::`) in order to disambiguate the `fuchsia` symbol.

Change-Id: I4a51df7ebe418addd146f448ad7a9a512cf2aa2b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663804
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Adam Lesinski <adamlesinski@google.com>