[system-update-committer] Completely disable in recovery images.
system-update-committer already has logic to skip the health check if
the current partition is a recovery partition. Unfortunately, if the
paver is unable to determine the current partition (e.g. because A/B/R
metadata is corrupted, or when using the installer image),
system-update-committer would immediately reboot the system.

This CL adds a "disable" config option which is false by default, and
includes it in recovery images.

Bug: 92116
Change-Id: I91ba61ed74838fc4cfabb6882950374adad4c493
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/652702
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Simon Shields <simonshields@google.com>