[ffx] Clean up ffx target when emu fails to start
The attached bug indicates a workflow where we were adding manual
targets to the ffx target collection, but failing to clean them up if we
encountered a later error.

This CL adds the clean-up step to that workflow, so that we don't leave
the manual targets behind.

Bug: fxbug.dev/95967
Change-Id: I27f90a3da208bb137278c9910b71be08f18eb785
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663506
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>