[wlan] Add pop_if to priority_queue
For convenience a user might need to remove frames from a priority queue
based on some attributes of those frames. As an example if a network
port goes down a driver might want to pop all frames from a queue that
belong to that port.

Add a method pop_if to priority_queue that takes a predicate that
evaluates each frame in the queue in priority order. If the predicate
returns true for a frame that frame will be popped.

Test: Included unit test
Change-Id: I3cdb3e8d46460c24ad2576aee42ee549d43f696b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/610863
Reviewed-by: Sakthi Vignesh Radhakrishnan <rsakthi@google.com>
Commit-Queue: Bjoern Johansson <bjoernj@google.com>