[ffx] Require an explicit flag when reusing emu
When an emulator instance is terminated but not cleaned up, the ffx
launcher can reuse the old instance's image files. This can be very
useful for certain workflows, but can also lead to some unexpected
behavior if the user doesn't know it's happening.

This CL adds a "--reuse" flag to the launcher. If the user attempts to
start an emulator with the same name as a persisted instance, it will
overwrite the existing images unless the "--reuse" flag is included on
the command line. This ensures that any reuse of old images is
intentional and expected.

Bug: fxbug.dev/94626
Change-Id: Ica32b2e1ab6d337590d6dacfef5d6680aa859276
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663115
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>