[sdk-dart][realm-builder] get/replaceComponentDecl, improve start/stop
This fixes an outdated test that now uses newly-enabled start event
handling to demonstrate late-binding to a child component (starting the
component after building the realm) and waiting for the component to start.

It also improves the event callback support to allow a test to subscribe
to more than one event (for example, both "started" and "stopped"), which
wasn't possible with the EventType-specific constructors.

Adds support for component decl manipulation, and demonstrates adding
the Binder protocol to a component that didn't expose it in CML.

Also adds support for get/replaceRealmDecl, and replaceConfigValue (just
via a FIDL ValueSpec for now). Tests for replaceConfigValue will need to
be added in a future CL.

Change-Id: If0c44ca5c687bb2d7b11e7aeaefe85d9a928cfc8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663803
Reviewed-by: Chase Latta <chaselatta@google.com>
Fuchsia-Auto-Submit: Rich Kadel <richkadel@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>