[starnix][signals] Move SignalActions to ThreadGroup
On Posix, signal handler are associated to the process and not the
thread. This CL reflects this by moving the ownership of the signal
handler table to the ThreadGroup. There is no behaviour change, as
previously, the code was making sure to correctly share the signal
handler table.

Change-Id: I56cd56a805469a81ec5e809524109fef06af42b8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663726
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>