[network/tests] Bind to address before client connects
The test currently involves a potential race between the server binding
to the address and the client attempting to connect to that address; if
the client wins, it'll see a "connection refused" error. Instead, the
server should bind to the address before allowing the client to connect.

Fixed: 96597
Multiply: netstack-forwarding-integration-test
Change-Id: Iab3302a6dc44f11cf3625f49292c266193646114
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/664284
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>