[ffx] Fix assumptions in the emu config flags
The attached bug indicates a workflow where the current QEMU flags
failed. This is due to unstated assumptions that were made for previous
versions of QEMU that no longer hold in the current version. The result
was that what we thought were defaults actually aren't, and so emulator
start-up failed in certain circumstances.

This CL adjusts the flags needed when running with --headless. This has
been tested on FEMU and QEMU5, over ssh (no display) and CRD (with and
without --headless). The result is that FEMU behavior is unchanged, but
QEMU5 no longer fails over ssh.

An additional side effect is that QEMU5 does not open a window with the
monitor console anymore, but a user can still access the monitor console
by passing the --monitor flag during start-up.

Bug: fxbug.dev/95860
Change-Id: I8fbaba1eb3ea48a865f89eb394ba141f804c4bd5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663960
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>