[credential_manager] Implements Add, Remove & Check Credential.
This change implements the Add, Remove and Check Credential operations for the
CredentialManager. The CredentialManager acts as the storage for the cr50
holding the HashTree and the encrypted CredentialMetadata.

On launch the CredentialManager checks if a HashTree already has been
persisted to disk and if so loads it into memory. Otherwise it will create
and persist a fresh HashTree and call ResetTree on the CR50.

Once the HashTree has been provisioned the CredentialManager starts
serving the fuchsia.identity.credential.CredentialManager protocol with
the three new operations.

Operations:
- AddCredential: Allocates a new credential in the hash tree binding a
  low-entropy secret to a high entropy secret in the CR50.
- CheckCredential: Validates an existing credential, if valid returns
  the high entropy secret.
- RemoveCredential: Removes an existing credential from the system.

Testing:
The vast bulk of this CL is test code testing the interactions of the
credential manager with a mocked pinweaver protocol.

Fixed: 92205
Fixed: 89058
Change-Id: I4b6b61a1c794b935a43d3a7df6eb0bedac7e5244
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/634604
Reviewed-by: Jody Sankey <jsankey@google.com>
API-Review: Jody Sankey <jsankey@google.com>
Commit-Queue: Ben Wright <benwright@google.com>