[fidlc] Sort declarations before emitting
This CL makes fidlc sort declarations by name before emitting them in
the JSON IR or coding tables. Previously, the ordering was determined by
the .fidl source order (and order of files given on the command line).

Motivation: my changes in I3ef8929bcc8bf0cc02f34c55e45e62f65123616d, the
FIDL Versioning CL, ended up having this effect. However, I believe it
to be a beneficial change, so rather than avoiding it, I am doing it
first in this CL explicitly. This will avoid a goldens diff on the
Versioning CL, which is important because it is a huge change and we
want to be sure it doesn't break anything.

Test: fx test fidlc_golden_tests
Bug: 67858 (FIDL versioning)
Change-Id: I30c2bc012e1d3454bcf35145fe7336372635aa5a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663931
Reviewed-by: Yifei Teng <yifeit@google.com>
Fuchsia-Auto-Submit: Mitchell Kember <mkember@google.com>
Reviewed-by: Pascal Perez <pascallouis@google.com>
Commit-Queue: Mitchell Kember <mkember@google.com>