[starnix] Add galaxies
This creates 4 different galaxies in which the starnix runner
operates. Each galaxy contains its own configuration and system image.
Each galaxy is exposed as a runner. For example, the "android" galaxy is
exposed as the "stardroid" runner, and the "android test" galaxy is
exposed as the "stardroid_test_runner".

This also changes the component hieararchy of the test runners. More
details in the README.

Change-Id: I4c909ca6e0dc3ea5c7cfbe1b3d4fa1442ca386d4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/662283
Reviewed-by: Ben Wright <benwright@google.com>
Reviewed-by: Adam Lesinski <adamlesinski@google.com>
Commit-Queue: Kevin Lindkvist <lindkvist@google.com>