[omaha-client] Soften opt-out logspam
fuchsia.update.config.OptOut is a protocol provided by some products and
not others. When connecting to a protocol that has no provider,
component manager currently logs a warning indicating it could not route
the connection to a provider, which looks scary on products that simply
do not have this component.

This change adds more logging to omaha-client when trying to use this
protocol to indicate that the protocol is not expected to exist on all
products in an effort to avoid that warning being a red herring in
investigations.

Eventually, CF will allow protocols to be marked as optional, silencing
this warning entirely.

Fixed: 94310
Change-Id: I94152e2c65e196a86f920522ca60e55974f95ae4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/664585
Reviewed-by: John Wittrock <wittrock@google.com>
Commit-Queue: Kevin Wells <kevinwells@google.com>