[bt][gatt] Add UpdateConnectedPeers method to gatt:GATT
Add a separate "UpdateConnectedPeers" method to gatt::GATT. While this
method could ostensibly be combined into UpdatePeer in a number of ways
(e.g. UpdatePeer could take an optional<PeerId> instead, or take a
vector<PeerId> which could be empty), the semantics around indication
confirmation callbacks are sufficiently different/confusing for this to
be a separate method (ex: for the vector<PeerId> case, is the callback
notified for each indication, or when all of them are complete? If the
former, how will the client know when all indications are complete for
the empty vector without knowledge of all current connections? If the
latter, what's the advantage of that vs. separate methods?).

Fixed: 657
Bug: 36375
Test: Moved arbitrary service registration into GattTest helper method
      for reuse. Added GattTest.{
         NotifyConnectedPeersNoneConnectedDoesntCrash,
         NotifyConnectedPeerWithConnectionDoesntCrash }
      Added GattIndicateMultipleConnectedPeersTest harness as a
      specialization of GattTest for testing indication callback
      semantics.
Change-Id: Ibf53a82e95b494a7a3520ff4d44d037ac45fd118
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/647003
Reviewed-by: Xo Wang <xow@google.com>
Commit-Queue: Lucas Jenkins <lucasjenkins@google.com>
Fuchsia-Auto-Submit: Lucas Jenkins <lucasjenkins@google.com>