[wlan] Add feature query methods to softmac AP.
The shift to using feature query methods will require the softmac AP
MLME to implement the query methods. This change implements them and
adds tests to verify that they work. A new test is added to verify
that the QueryDeviceInfo method works as well.

Bug: 88315
Test: fx test wlan-mlme wlan-hw-sim wlanstack

Change-Id: I03ec51fdca27dc7a9bf8b11603f60ab944b8e3ab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663382
Reviewed-by: Dylan Swiggett <swiggett@google.com>
Commit-Queue: Karl Ward <karlward@google.com>