[stress_tests] Fix broken Component Manager stress test
Component Manager stress tests do not currently run on CQ. A
recent change to the open flags in fuchsia.io broke the compilation
of this test, because it was not updated.

Fixed the broken test.

Added a new CQ variant of this test that should run for a minute
before terminating.

Test: fx test component-manager-stress-tests-alt
      fx test component-manager-stress-tests-short
Change-Id: I8ba3ff9f630469ba24f6ef0182cb8764749d53c7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/664682
Reviewed-by: Tamir Duberstein <tamird@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Xyan Bhatnagar <xbhatnag@google.com>