[bt][hfp] Send call held status indicator on swap.
When  we have an active call and a call held, and we get a AT+CHLD=2 (hold
active message), HFP should swap the two calls (by sending a fidl message
to the call manager) and send a +CIEV:4,1 (call held=1) indicator.  We were
not sending this indicator because there is no actual change to the call
state.  This change forces the calls stream to include the call held =
1 indicator if an active and held call are swapped, which causes the
  indicator AT message to be sent to the HF.

Tests: HFP/AG/TWC/BV-03-I
       fx test fuchsia-pkg://fuchsia.com/bt-hfp-audio-gateway-tests#meta/test-component.cm
Bug: http://fxb/90227
Change-Id: I89df030644f679ceb8fd0a8820e24b2b52b89df8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/658502
Reviewed-by: Ani Ramakrishnan <aniramakri@google.com>
Reviewed-by: Marie Janssen <jamuraa@google.com>
Commit-Queue: Matthew Kehrt <kehrt@google.com>