[cr50_agent] Convert nanos to seconds
A `zx_duration_t` in FIDL is represented as an i64 of nanoseconds.  The
cr50 expects a u32 of whole seconds.  We need to actually convert from
one representation to the other or we'll fail to enroll a credential or
worse, lock users out for a billion times as long as we meant to.

Change-Id: I65b93dde716075e114dbfeba81b13bc018cd2584
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/664982
Reviewed-by: Simon Shields <simonshields@google.com>
Commit-Queue: Drew Fisher <zarvox@google.com>