[netstack3] Add receiving dev for packet handling
For SO_BINDTODEVICE, the UDP packet handler needs to be aware of what
device a packet was received on. This CL threads that information
through the stack so that it is available at UDP socket lookup time,
though it is currently unused.

Bug: 96573
Change-Id: I248a595c279aa3bdfbc8c0410abaea57fe6d6391
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/664697
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Alex Konradi <akonradi@google.com>
Fuchsia-Auto-Submit: Alex Konradi <akonradi@google.com>