[starnix][pts] Fix pts node lifecycle
The pts node must be tied to the main terminal being opened and not to
the terminal object being lived. In particular, it must be removed when
the main terminal fd is closed even if some replica are still opened.

The previous code had test passing because the terminal object was not
owned by the replica file descriptors. This is changed in this CL and
the correct lifecycle is ensured.

Change-Id: I5eb44c0621ab9f4173e3eaf5266b72febc0451c2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/665065
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Benjamin Lerman <qsr@google.com>