[test_manager] Decouple ArchiveAccessor in mocks and client logs
Removes the ArchiveAccessor proxy shared between the mock component
and held by test_manager for serving client logs. We are seeing some
odd hangs that are related to logs and seem to be correlated with tests
that access their own diagnostics (and therefore would exercise the
mock). This simplifies the routing so that the mock connects via the
handles it is given, more like a "real" component.
The goal is to simplify the diagnostics system somewhat to either
eliminate or better understand the hangs.

Note: this also moves the mock within the topology somewhat to avoid a
dependency cycle.

Testing: fx test //src/sys/test_manager, test coverage in CI
Bug: 94333

Change-Id: Idc642235f44d96324f2c2ae2641f2dcd418a3f70
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/664066
Reviewed-by: Ankur Mittal <anmittal@google.com>
Commit-Queue: Satsuki Ueno <satsukiu@google.com>