[bt][fastpair] Define a Model ID type
The Model ID is assigned to a device during registration. This
varies based on the device and is fixed unless the device is
unregistered in the Nearby Devices Console.

Define a local type for the 24-bit Model ID with conversions to/from
the relevant primitives. This will be used by the GATT service & LE
Advertiser.

Bug: 94166
Test: fx test bt-fastpair-provider-tests
Change-Id: I9de70873ae33162f668e63224363f21c5525a998
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/658789
Reviewed-by: Lucas Jenkins <lucasjenkins@google.com>
Fuchsia-Auto-Submit: Ani Ramakrishnan <aniramakri@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>