[fidlc] Warn on attribute typos earlier
This CL moves the attribute typo detection from VerifyAttributesStep to
ConsumeStep. This means these warnings now get reported even if there
are errors during the later compilation steps.

Motivation: most attributes are just extra information that we validate
and pass through to the IR, but there are some that affect compilation,
like @available and @generated_name. These can be load bearing for
correctness: if you remove it, or misspell it (making it a user-defined
attribute), the library fails to compile. An example:

    type Foo = struct {
        // This attribute is crucial to avoid a name collision with Foo.
        foo @generated_name("Bar") struct {};
    };

If you misspell @generated_name, compilation will fail due to the name
collision on "Foo". If you don't spot the typo, this will be confusing,
so this CL ensures we report the typo warning cases like these.

Test: fx test fidl-compiler
 Please enter the commit message for your changes. Lines starting
Change-Id: Id337fafb7c0e27d28c1b9350f01258e7cbb3d605
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/664802
Reviewed-by: Alex Zaslavsky <azaslavsky@google.com>
Commit-Queue: Mitchell Kember <mkember@google.com>