[debugger] Allow async thread controller stops.
This adds the capability to have a thread controller respond that it
wants to respond to the stop later. The same stop will be reissued when
the thread controller reports that it's done.

This is not currently used outside of the newly-added test. This will be
used in a future patch to continue conditionally whether a breakpoint
add has completed or not.

Bug: 95664

Change-Id: I68892ed8de39ee284070e9b69b556deedea68559
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/664883
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>