[netstack3] Perform route lookup without recursion
This is purely a mechanical change. No new feature is introduced and
the stack continues to only allow setting a local route or route through
a gateway without specifying a device (which https://fxbub.dev/96680
will eventually fix).

Note that this change has no performance impact in the lookup/send path
compared to the forwarding table implementation before
https://fuchsia.googlesource.com/fuchsia/+/8a849ee4b9abbfde582be0a7e0e3bd301368991f
which removed the concept of an "active" table. By removing the "active"
table, lookup was no longer O(1) (was O(n) because of the recursive
property on the installed table). This change removes the recursive
property so that lookup operations are O(1) again. Insertion is now
less expensive than the pre-"active" table implementation as we
no longer need to calculate an "active" table on each route table
change.

Fixed: 96162

Change-Id: I2ca556d9b51165e86b69a8213a803ebd48e6d3d5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/661726
Reviewed-by: Alex Konradi <akonradi@google.com>
Fuchsia-Auto-Submit: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Joshua Liebow-Feeser <joshlf@google.com>
Commit-Queue: Ghanan Gowripalan <ghanan@google.com>