[debugger] Allow stepping over syscalls.
This implementation detects failures of the "until" controller and falls
back to single-stepping through PLT trampolines. This handles the case
where the destination function is in read-only memory (like the vDSO).

The challenge is that this forces initialization of the until controller
(bubbling up to the Function step thread controller) to be asynchronous.
The previous design didn't allow this.

The extra trickiness is that sometimes initialization is synchronous and
sometimes asynchronous and the return value from OnThreadStop needs to
vary based on this. A new helper class is added to the ThreadController
to handle this complex state combination.

Bug: 95664
Change-Id: I189b2bba46dbc5ae4e308d3393f12b8c66c19323
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/664884
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>