[netstack3] Replace try_insert() with entry()
Some operations will require examining the current value in the map, and
then inserting if there is none or appending if there is one (and some
other conditions are satisfied). Without entry(), this would have to be
done with a try_insert() then apply(), which duplicates the number of
map lookups being performed.

Bug: 94652

Change-Id: I676079d4d3568facafd093db534596d82ad30997
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/660662
Reviewed-by: Joshua Liebow-Feeser <joshlf@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Alex Konradi <akonradi@google.com>