[virtualization] Enable WriteGoodAndBadSectors test for virtio_block_rs.
Update the test to allow the rust device to reject the entire chain if
any part of it is beyond the capacity of the device. We don't change the
behavior of the C++ component since we will likely be removing this
component before too long anyways.

Bug: 95529
Test: fx test device_tests
Change-Id: I110df8a96e3e61be2b92da5e0a454edf80191175
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663534
Reviewed-by: David Hastings <dahastin@google.com>
Commit-Queue: Tim Detwiler <tjdetwiler@google.com>