[zircon][libc] Use BREAK_ON_LOAD instead of r_brk_on_load
Use ZX_PROP_PROCESS_BREAK_ON_LOAD to store the address of the debug trap
that indicates a change of the dso list. In the past the value is stored
in r_debug.r_brk_on_load, but it's not portable to starnix processes
with other libcs.

This property is not used anywhere yet except in the tests, so existing
behaviors should not change.

Bug: 92960
Change-Id: I1ccabfc81a9d641a6b47beec9f3c22a2af4d7fd7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/652543
Reviewed-by: Carlos Pizano <cpu@google.com>
API-Review: Carlos Pizano <cpu@google.com>
Reviewed-by: Brett Wilson <brettw@google.com>
Commit-Queue: Dangyi Liu <dangyi@google.com>