[core][zbi_test] Add timeout to the hpet test on asan.
The test currently takes around 6 min which is the default infra
timeout per test, thus causing the test to flake frequently with time
outs. We should explicitly set the timeout to a larger timeout for this
test, so that infra can respect the new timeout.

Bug: 95826
Change-Id: I2ee36d193b56493cd807888af4ed2fcecadf090b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/665609
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Ina Huh <ihuh@google.com>