[mdns] add ProxyHostPublisher and ServiceInstancePublisher protocols
This CL adds ProxyHostPublisher and ServiceInstancePublisher protocols
to the mdns service, the latter being a replacement for Publisher.
Quite a bit of new plumbing was required to deal with more specific
transport interaction such as restricting an agent to a particular
medium (wired or wireless) and IP version.

Test: tests added a existing tests updated

Bug: 93994,93320,94918

Change-Id: I666945fa42e9d082b4f61fc16da7e20e9c8267f4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/656962
Reviewed-by: Robert Quattlebaum <rquattle@google.com>
API-Review: Dale Sather <dalesat@google.com>
Reviewed-by: Nate Hurley <nahurley@google.com>
Commit-Queue: Dale Sather <dalesat@google.com>