[fshost] Route core the delayed pkgfs directories
Core is stored in pkgfs, so the delayed directories are guaranteed to
be serving before core or anything below it interacts with them.

Also, to highlight the unusual behavior of the not-delayed directories,
add the "-not-delayed" suffix to the not-delayed directories and remove
the "-delayed" suffix from the delayed directories.

Change-Id: I810d988042f2bf312deec06fbc3c916e84d37aaf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663684
Reviewed-by: Stephen Demos <sdemos@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Allison Pearce <ampearce@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Commit-Queue: Ben Keller <galbanum@google.com>