[driver_host] Serialize all calls into driver via driver runtime
This will enable us to remove device_get_dispatcher. Now that every
driver has a dispatcher, we can serialize all calls into drivers via
these dispatchers. This does induce a thread hop for call originating
from the main driver host thread, as well as a thread hop when calling
from one driver into another.

A future change can improve the performance of FIDL handeling by
creating a per driver instance vfs which hosts the devfs vnode. That vfs
can be use the per driver dispatcher. This will alleviate the need to
hop threads when invoking the message hook.

One of the downsides of this implementation is that we will spawn a
thread per driver instance. This is better than the current state where
we spawn a thread per device. In order to address this problem we would
need to make driver host APIs async, but that is probably too difficult
to do realistically.

Re-entrancy in the driver host had be to cleaned up. This does change
behavior slightly in that device_add will return before calling the open
hook on the device, but hopefully no one depends on that behavior.

Lastly, closure-queue, aml-securemem, and virtual-audio were updated to
not use the current thread as a proxy for asserting they are
synchronized. Instead, where possible, asserting that the current
dispatcher is the same as the stored dispatcher is perfomed instead.

Tested: fx test driver-host-tests
Change-Id: I3b2b59ab954dee69c2cb4b9a9a135f8a29919759
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/664506
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>