[ffx] ffx component doctor command v0
First iteration for the new `ffx component
doctor` command. Shows outgoing capabilities
that are not exposed and vice-versa.

`exposed` capabilities are declared in the
	manifest.
`outgoing` capabilities are available at
	runtime.

This check is used to verify that the `exposed`
capabilities are available at runtime and that
there are no extra runtime capabilities not
declared in the manifest.

Bug: 94058
Change-Id: I1f3ea658c33adc1371e8384743b56120154ff044
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/654143
Reviewed-by: Adam Perry <adamperry@google.com>
Reviewed-by: Xyan Bhatnagar <xbhatnag@google.com>
Commit-Queue: Simon Tran <transimon@google.com>