[fdf][compat] Fake sysmem fragment.
Many DFv1 drivers expect to have a sysmem fragment. This is something we
want to move away from in DFv2 -- drivers should have sysmem routed
directly to them.

To make this simpler in the interim, we add support for a fake sysmem
fragment to the compat driver.

Bug: 91608
Change-Id: I61ebc95091daba23f937f7ed2c379ef0ed1f2867
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/657042
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Simon Shields <simonshields@google.com>