[netstack3] Support enabling/disabling an IP device
This paves the way to hold all device state in core instead of sharing
it with bindings when the device is disabled (we currently remove it
from core and then re-insert when the device is re-enabled).

This change properly handles starting/stopping the following when the
device is enabled/disabled:
- IPv6 Router Solicitation
- DAD

A later change will properly filter routes by device enabled
configuration.

Bug: 92656

Change-Id: If88f1c40d2359d2b9a3672e188831392ae4415ea
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/664849
Reviewed-by: Alex Konradi <akonradi@google.com>
Fuchsia-Auto-Submit: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Joshua Liebow-Feeser <joshlf@google.com>
Commit-Queue: Ghanan Gowripalan <ghanan@google.com>