[netemul] Wait for test to exit before cleaning up networks
Currently, the netemul runner signals that the test component has
finished running once the test is complete, and then cleans up the
virtual network environment it created for the test. This can cause a
race between network cleanup and realm shutdown: if a component in the
test realm is using one of the virtual endpoints, even though the test
itself is complete, it might log an error or panic if the endpoint is
removed out from under it.

To address this, after closing the component controller channel, the
test runner waits to observe the stopped event from the test root; once
the root is stopped, it can assume that all the constituent components
have been shutdown in an orderly fashion as well. Then it's safe to drop
the network environment.

Multiply: netemul-runner-tests
Multiply: netemul-runner-errors-tests
Change-Id: Id9265355e1ac84268c070e9051eb6bd1580b7566
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/666244
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>