[component][fuzzing] Clean up ComponentContext
This CL modifies ComponentContext to allow constructing a client-only
version that does not consume the directory channel from the startup
handles. When adding the target instrumentation library to a binary to
be fuzzed, it must not consume this handle, or the process will not be
able to serve FIDL protocols of its own.

The CL also refactors the remaining places that were calling
sys::ComponentContext directly with calls to fuzzing::ComponentContext.

Change-Id: I77c16f312d874fbd66bd3414e62c68af5d7bf9e8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/662576
Reviewed-by: Nick Cano <nickcano@google.com>
Commit-Queue: Aaron Green <aarongreen@google.com>