[ftl] Delete read/write fuzzer.
This fuzzer mostly wrote random data to the device and ensured that it
could read the same set of random data back. As a fuzzer it is
essentially just testing the same operation repeatedly with different
data, but that data is not actually processed, it is treated as an
opaque blob. While this not being true would be worrying, it is not a
real issue that is expected to ever come up, and if it does then we've
moved on to doing something very strange.

Fixed: 94146

Change-Id: I8bcc5257493827b744e1acd9be19ce84be5eafa8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/666599
Reviewed-by: Brandon Castellano <bcastell@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>