[kernel][arm64] remove FIQ support
There was a residual FIQ support in the exception handling code that has
never been used and probably not likely to in the foreseeable future.
FIQs are usually subsumed by secure mode interrupts and unavailable to
EL1 guests.

It was always stubbed out as unimplemented anyway, so this simply cleans
up a bit of code.

Change-Id: I518c40693ebcb218a2dc3d94bcae61820ed51909
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663927
Reviewed-by: Corey Tabaka <eieio@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Travis Geiselbrecht <travisg@google.com>