[fdf] Add inequality and value lists to device groups
Update the device_group.fidl so that it supports
a property list and conditions for the device groups.

Modify the logic in driver index so that it evaluates
the conditions and value list.

Bug: 96714
Tests: Added unit tests in driver-index-unittests
Change-Id: Ic9a3c7cc13af5824f5cf44da1914cafbe5275b1f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663862
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
API-Review: Jocelyn Dang <jocelyndang@google.com>
Commit-Queue: Sarah Chan <spqchan@google.com>