[ffx][pkg] Fall back to use root.json if 1.root.json is not available
Some repositories start with a non-zero root version and expect the client to handle this situation. Unfortunately ffx repository doesn't really support the user to provide this information. As a stopgap, this changes ffx repository to fallback to try the non-versioned `root.json`, which is not formally part of the TUF spec, but most repositories create it, and it corresponds to the latest root metadata.

Bug: 92126
Change-Id: I8fcb789837f3c8c0cadba9a6b88fe78ad7538ca1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/665943
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Jiaming Li <lijiaming@google.com>