[insntrace] Add "flaky" label for insntrace_tests on NUC
The test is flaky on NUC and blocks the builder. We are disabling the
test on CQ builders until the flakiness issue is solved. The test will
be only run on FYI builders for flaky tests.

Also the test is not supported on FEMU, so we also removed the "emu_env"
from the environments list.

Bug: 97097
Change-Id: Ieb2b0135d0e09096779ca0a5461f2128eba26c6e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/666743
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Reviewed-by: Marc-Antoine Ruel <maruel@google.com>
Fuchsia-Auto-Submit: Yilong Li <liyl@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>