[kernel][sched] Fix cluster assignment when favoring LITTLE cores.
Fix an oversight when re-ordering the CPU search sets to favor LITTLE
cores over big cores. The CpuSearchSet::cluster() accessor assumed that
entry 0 is always the CPU the search set is associated with, however,
when ordering the CPUs to favor little cores this is no longer true, as
LITTLE cores should be the first entries for all search sets.

- Update the accessor to use the global cluster map instead of the 0th
  entry in the search set to fix the now incorrect assumption.
- Reduce the cross-cluster migration threshold, which is now being
  correctly applied in heterogeneous systems.

Associating this CL with the power-aware load balancing tracking bug for
context in that effort.

Bug: 96677
Change-Id: I2a9d718dd7d701463db00aec9f6988a9fea57894
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/666848
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Corey Tabaka <eieio@google.com>